another gcc/clang fix

+++ CesiumGeospatial/src/S2CellBoundingVolume.cpp
@@ -44,7 +44,7 @@ std::array<Plane, 6> computeBoundingPlanes(
   glm::dvec3 verticesCartesian[4];
 
   double maxDistance = 0;
-  for (int i = 0; i < 4; ++i) {
+  for (size_t i = 0; i < 4; ++i) {
     verticesCartographic[i].height = s2Cell.getMinimumHeight();
     verticesCartesian[i] =
         ellipsoid.cartographicToCartesian(verticesCartographic[i]);
